+++ src/main/Enum/ErrorCode.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace WebArch\BitrixTaxidermist\Enum;
+
+class ErrorCode
+{
+    const ERROR_REGISTERING_AUTOLOADER = 1;
+
+    const AUTOLOADER_IS_ALREADY_REGISTERED = 2;
+
+    const AUTOLOADER_NOT_REGISTERED = 3;
+
+}