+++ src/main/Exception/TaxidermistExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace WebArch\BitrixTaxidermist\Exception;
+
+interface TaxidermistExceptionInterface
+{
+}